refactor: sampling into its own file
